[ci skip] first pass, pluggable label selection
