+++ lib/copy.js
@@ -1,7 +1,6 @@
 const tokenize = require('./tokenize');
 const tokens = require('@mapbox/geocoder-abbreviations');
 const diacritics = require('diacritics').remove;
-const title = require('./titlecase');
 const readline = require('readline');
 const turf = require('@turf/turf');
 const fs = require('fs');
@@ -31,12 +30,12 @@ function start(callback) {
 
     rl.on('close', () => {
         if (!opts.solo) {
-	    opts.output.end(() => {
-		process.send(unitMap);
-	    });
-	}
-	if (callback && (typeof callback === 'function'))
-	    return callback();
+            opts.output.end(() => {
+                process.send(unitMap);
+            });
+        }
+        if (callback && (typeof callback === 'function'))
+            return callback();
     });
 
     rl.on('line', (data) => {
@@ -84,7 +83,8 @@ function start(callback) {
             feat.properties.street = '';
         }
 
-        feat.properties._text = title(feat.properties.street);                              //The _text value is what is displayed to the user - it should not be modified after this
+        // the feature _text value that is displayed to users is determined later, in map.js w/ a pluggable lib/labels/*.js system
+        feat.properties._text = feat.properties.street;
 
         let tokens = tokenize.main(feat.properties.street, opts.tokens, true);
         feat.properties.street = diacritics(tokens.tokens.join(' '));                       //The street is standardized and it what is used to compare to the address cluster
