+++ lib/map.js
@@ -29,6 +29,7 @@ module.exports = function(argv, cb) {
                 'output',
                 'in-network',
                 'in-address',
+                'label',
                 'map-network',
                 'map-address',
                 'error-network',
@@ -48,6 +49,7 @@ module.exports = function(argv, cb) {
                 'map-address': 'map-addresses',
                 'segments': 'segment',
                 'database': 'db',
+                'label': 'l',
                 'output': 'o',
                 'tokens': 'token'
             }
@@ -73,7 +75,7 @@ module.exports = function(argv, cb) {
     if (argv.segment) argv.segment = path.resolve(__dirname, '..',  argv.segment);
 
     if (argv.tokens)
-	argv.tokens = tokenize.createReplacer(argv.tokens.split(','));
+        argv.tokens = tokenize.createReplacer(argv.tokens.split(','));
 
     const output = fs.createWriteStream(path.resolve(__dirname, '..', argv.output));
 
@@ -378,6 +380,12 @@ module.exports = function(argv, cb) {
     }
 
     function orphanAddr() {
+        let label;
+        if (!argv.label)
+            label = require('./lib/label/titlecase')();
+        else
+            label = require('./lib/label/' + argv.label)();
+
         pool.connect((err, client, pg_done) => {
             if (err) return cb(err);
 
@@ -391,7 +399,7 @@ module.exports = function(argv, cb) {
                     total: parseInt(res.rows[0].count)
                 });
 
-                const cursor = client.query(new Cursor(`SELECT a._text, ST_AsGeoJSON(a.geom) AS geom FROM address_cluster a LEFT JOIN network_cluster n ON a.id = n.address WHERE n.address IS NULL;`));
+                const cursor = client.query(new Cursor(`SELECT a._text AS address_text, n._text as network_text, ST_AsGeoJSON(a.geom) AS geom FROM address_cluster a LEFT JOIN network_cluster n ON a.id = n.address WHERE n.address IS NULL;`));
 
                 return iterate();
 
@@ -403,16 +411,16 @@ module.exports = function(argv, cb) {
                             pg_done();
                             console.error('ok - addresses assigned to linestring');
                             return cluster.adoption(() => {
-				console.error('ok - adopted stranded address clusters into stable homes');
-				cluster.prune(finalize);
-			    });
+                                console.error('ok - adopted stranded address clusters into stable homes');
+                                cluster.prune(finalize);
+                            });
                         }
 
                         rows.forEach((row) => {
                             let feat = {
                                 type: 'Feature',
                                 properties: {
-                                    'carmen:text': row._text,
+                                    'carmen:text': label(row),
                                     'carmen:center': false,
                                     'carmen:addressnumber': []
                                 },
@@ -458,9 +466,9 @@ module.exports = function(argv, cb) {
 
     function finalize(err) {
         if (err)
-        return cb(err);
+            return cb(err);
 
-	console.error('ok - pruned multiple network_clusters matched to the same address_cluster');
+        console.error('ok - pruned multiple network_clusters matched to the same address_cluster');
 
         if (argv['error-network']) argv['error-network'].close();
         if (argv['error-address']) argv['error-address'].close();
